Convert emoji.
